+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Update to newer provider resolver style
 - Update to new repo key used for Ubuntu/Debian installs
 - Bump apt dependency from 2.x to 3.x
+- Forcibly install the xdg-utils dependency on Debian platforms
 
 v0.1.0 (2015-05-26)
 -------------------